Add files via upload
The program is set to grab and reference the "Maze.jpg" from its location on my computer, and write the changes to and output the "Output.jpg" file to a location on my computer. To change this, in the "PossiblePath.java" class change the file "Maze.jpg"'s path so that it is set to the location of the file on your computer, and likewise with the path of the "Output.jpg" file. Ps. The Output file currently holds the solution to the maze (2D array) I have declared as "maze" in "PossiblePath.java", if you choose to make your own maze, then that maze's solution will override the current solution in the "Output.jpg" file.